+++ NEWS.md
@@ -4,3 +4,4 @@
 * Added verbosity to `list_surveys()`
 * Added `get_citation()` - Resolves #38
 * Added bibtex citation style as default to `get_citation()` - Resolves #52
+* Fixed verbosity issue by importing purrr::quietly - Resolves #68